requirements.txt
pillow>=10.0.0
edge-tts>=6.1.0
